\name{Numerical Approximation of the Force of Infection}
\alias{foi.num}
\title{Numerical Approximation of the Force of Infection}
\usage{
foi.num(x, p)
}
\arguments{
\item{x}{A numeric vector depicting age categories.}
\item{p}{A numeric vector depicting the seroprevalence.}
}
\description{
Numerical approximation of the force of infection based on age and estimated seroprevalence.
}
\seealso{\code{\link{mpspline.fit}}}
