\name{HCV_BE_06}
\alias{HCV_BE_06}
\docType{data}
\title{
Hepatitis C Serological Survey Data in IDUs (Belgium, 2006)
}
\description{
The Hepatitis C virus (HCV) is the leading cause of known liver disease in most industrialized countries. Mathei et al. (2006) presented a study of HCV infection among injected drug users (IDUs). The data consists of IDUs from three areas in Belgium: Charleroi, Antwerp, and the province of Limburg.
}
\usage{data("HCV_BE_06")}
\format{
  A data frame with 25 observations on the following 7 variables.
  \describe{
    \item{\code{V1}}{A numeric vector containing the duration of injection (in years).}
    \item{\code{V2}}{A numeric vector containing the amount of users.}
    \item{\code{V3}}{A numeric vector containing the amount of users (same vector as V2).}
    \item{\code{V4}}{A numeric vector containing the seroprevalence of HCV by duration of injection.}
    \item{\code{V5}}{A numeric vector (unknown function).}
    \item{\code{V6}}{A numeric vector (unknown function).}
    \item{\code{V7}}{A numeric vector (unknown function).}
  }
}
\details{
See page 65 of the book.
}
\source{
\url{http://www.ncbi.nlm.nih.gov/pubmed/16901287}
}
\examples{
data(HCV_BE_06)
}
\keyword{datasets}
