\name{Social Contact Data Matrices - Location}
\alias{loc_home}
\alias{loc_school}
\alias{loc_work}
\alias{loc_leisure}
\alias{loc_transport}
\alias{loc_otherplace}
\docType{data}
\title{
Social Contact Data Matrices - Location
}
\description{
The book works with six different social contact data matrices that consider the location of contact.
}
\usage{data("loc_home")
data("loc_school")
data("loc_work")
data("loc_leisure")
data("loc_transport")
data("loc_otherplace")}
\format{
  A data frame depicting a social contact matrix, with different stratification levels.
}
\details{
See page 237 in the book.
}
\examples{
data("loc_home")
data("loc_school")
data("loc_work")
data("loc_leisure")
data("loc_transport")
data("loc_otherplace")
}
\keyword{datasets}
