\name{Rubella_UK_8687}
\alias{Rubella_UK_8687}
\docType{data}
\title{
Rubella Serological Survey Data (UK, 1986-1987)
}
\description{
Rubella is a childhood disease that occurs worldwide. Commonly known as German measles, rubella usually causes a mild rash, sore throat, and swollen glands. The symptoms, however, are more severe in adults. The prevalence of rubella in the UK was obtained from a large survey conducted in the UK, anno 1986-1987 (Morgan-Capner et al., 1988).
}
\usage{data("Rubella_UK_8687")}
\format{
  A data frame with 44 observations on the following 3 variables.
  \describe{
    \item{\code{Age}}{A numeric vector containing age.}
    \item{\code{Pos}}{A numeric vector containing the number of seropositives in each age group.}
    \item{\code{Neg}}{A numeric vector containing the number of seronegatives in each age group.}
  }
}
\details{
See page 69 of the book.
}
\source{
\url{http://www.ncbi.nlm.nih.gov/pubmed/3142541}
}
\examples{
data(Rubella_UK_8687)
}
\keyword{datasets}
\seealso{\code{\link{Rubella_Mumps_UK_8687}}, \code{\link{Mumps_UK_8687}}}
