\name{Social Contact Data Matrices}
\alias{scd_all_contacts}
\alias{scd_close_m15}
\alias{scd_close_nonclose_p1h}
\alias{scd_close_p15_nonclose_p1h}
\alias{scd_close_p15}
\alias{scd_close_p1h}
\alias{scd_close_p4h}
\alias{scd_close}
\alias{scd_nonclose_p1h}
\alias{scd_nonclose}
\docType{data}
\title{
Social Contact Data Matrices
}
\description{
The book works with ten different social contact data matrices.
}
\usage{data("scd_all_contacts")
data("scd_close_m15")
data("scd_close_nonclose_p1h")
data("scd_close_p15_nonclose_p1h")
data("scd_close_p15")
data("scd_close_p1h")
data("scd_close_p4h")
data("scd_close")
data("scd_nonclose_p1h")
data("scd_nonclose")}
\format{
  A data frame depicting a social contact matrix.
}
\details{
See page 236 in the book.
}
\examples{
data("scd_all_contacts")
data("scd_close_m15")
data("scd_close_nonclose_p1h")
data("scd_close_p15_nonclose_p1h")
data("scd_close_p15")
data("scd_close_p1h")
data("scd_close_p4h")
data("scd_close")
data("scd_nonclose_p1h")
data("scd_nonclose")
}
\keyword{datasets}
